refactor: convert theme-toggle to TypeScript module and split tests
- Move theme-toggle from public/ to src/components/theme-toggle.ts as proper ES module
- Add build script using esbuild to compile TS to public/theme-toggle.js
- Update package.json to use ES modules and add js:build script
- Split test files into organized structure:
  - test/components/ for theme-toggle unit, init, and integration tests
  - test/worker/ for package routes, list, and UI element tests
- Update vitest config to track src/ coverage and new test paths
- Rename postcss.config.js to .cjs for ES module compatibility
- Add public/theme-toggle.js to .gitignore as build artifact

Tests remain comprehensive (27 passing) with 100% coverage on theme-toggle module.
